Moved the core parameters from the wx gui to ConfigManager

+++ src/common/ConfigManager.cpp
@@ -241,7 +241,7 @@ int* rewindSerials = NULL;
 u32 autoFrameSkipLastTime;
 u32 movieLastJoypad;
 u32 movieNextJoypad;
-unsigned short throttle;
+int throttle;
 
 const char* preparedCheatCodes[MAX_CHEATS];
